let git-fs use a commit instead of the index to get its files
Right now, this is hardcoded to "HEAD", but it will be configurable in
the future.

This change should also improve the performance, since a tree object is
actually stored in a tree datastructure and the lookup at each level
uses a binary search as well (instead of the single linear search
through all the filenames that was done before).

As an extra advantage, a tree contains explicit (subtree) objects for
directories, which removes the need to fiddle with prefix matching.

This commit also removes a lot of duplicate lookups by locating the
right tree entry in the open / opendir call and then using that pointer
in the read call.

Finally, it switches readdir from writing back its results all at once
(like before, which needs fuse to allocate additional memory for storing
those objects) to just filling up one buffer worth of filenames and
letting fuse call a second time.