Process the git repo path using realpath()
When not running on the foreground, fuse does a chdir("/") as part of
the detaching process, which breaks any relative repository path given
(since we don't use it until after fuse_main is done initializing). By
turning the path into an absolute path during argument processing, it
will remain working.