Revert "No need to init&shutdown threads anymore https://developer.gnome.org/glib/stable/glib-Deprecated-Thread-APIs.html"
This reverts commit 4fdf5df4b9af69c3ae558057564cefd4d72d52bc.

That commit linked to deprecated glib thread functions, and then went ahead and
removed libgit2 thread functions. The latter are very much not deprecated and
required to prevent random crashes and other problems. Note that they always
need to be called when libgit2 is compiled with thread safety enabled, even
when actually running singlethreaded.