Print errors on stderr, even after fuse initialization
During intitialization (when --debug is not passed), fuse_main daemonizes and
redirects stdin, stdout and stderr to /dev/null. This means any error are no
longer shown. This hides errors during execution, but also errors in the second
half of initialization.

This duplicates the stderr file descriptor and keeps it open, allowing errors
and backtraces to be printed later on.